requirements.txt
phonenumbers~=8.13.55
python-dotenv~=1.0.1
aiogram~=3.18.0
python-docx~=1.1.2
pandas~=2.3.0